+++ projects/hipcub/hipcub/include/hipcub/backend/rocprim/device/device_run_length_encode.hpp
@@ -1,7 +1,7 @@
 /******************************************************************************
  * Copyright (c) 2010-2011, Duane Merrill.  All rights reserved.
  * Copyright (c) 2011-2018, NVIDIA CORPORATION.  All rights reserved.
- * Modifications Copyright (c) 2017-2025, Advanced Micro Devices, Inc.  All rights reserved.
+ * Modifications Copyright (c) 2017-2026, Advanced Micro Devices, Inc.  All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -43,15 +43,17 @@ public:
     template<typename InputIteratorT,
              typename UniqueOutputIteratorT,
              typename LengthsOutputIteratorT,
-             typename NumRunsOutputIteratorT>
-    HIPCUB_RUNTIME_FUNCTION static hipError_t Encode(void*                  d_temp_storage,
-                                                     size_t&                temp_storage_bytes,
-                                                     InputIteratorT         d_in,
-                                                     UniqueOutputIteratorT  d_unique_out,
-                                                     LengthsOutputIteratorT d_counts_out,
-                                                     NumRunsOutputIteratorT d_num_runs_out,
-                                                     int                    num_items,
-                                                     hipStream_t            stream = 0)
+             typename NumRunsOutputIteratorT,
+             typename NumItemsT>
+    HIPCUB_RUNTIME_FUNCTION
+    static hipError_t Encode(void*                  d_temp_storage,
+                             size_t&                temp_storage_bytes,
+                             InputIteratorT         d_in,
+                             UniqueOutputIteratorT  d_unique_out,
+                             LengthsOutputIteratorT d_counts_out,
+                             NumRunsOutputIteratorT d_num_runs_out,
+                             NumItemsT              num_items,
+                             hipStream_t            stream = 0)
     {
         return ::rocprim::run_length_encode(d_temp_storage,
                                             temp_storage_bytes,
@@ -67,17 +69,18 @@ public:
     template<typename InputIteratorT,
              typename UniqueOutputIteratorT,
              typename LengthsOutputIteratorT,
-             typename NumRunsOutputIteratorT>
-    HIPCUB_DETAIL_DEPRECATED_DEBUG_SYNCHRONOUS HIPCUB_RUNTIME_FUNCTION static hipError_t
-        Encode(void*                  d_temp_storage,
-               size_t&                temp_storage_bytes,
-               InputIteratorT         d_in,
-               UniqueOutputIteratorT  d_unique_out,
-               LengthsOutputIteratorT d_counts_out,
-               NumRunsOutputIteratorT d_num_runs_out,
-               int                    num_items,
-               hipStream_t            stream,
-               bool                   debug_synchronous)
+             typename NumRunsOutputIteratorT,
+             typename NumItemsT>
+    HIPCUB_DETAIL_DEPRECATED_DEBUG_SYNCHRONOUS HIPCUB_RUNTIME_FUNCTION
+    static hipError_t Encode(void*                  d_temp_storage,
+                             size_t&                temp_storage_bytes,
+                             InputIteratorT         d_in,
+                             UniqueOutputIteratorT  d_unique_out,
+                             LengthsOutputIteratorT d_counts_out,
+                             NumRunsOutputIteratorT d_num_runs_out,
+                             NumItemsT              num_items,
+                             hipStream_t            stream,
+                             bool                   debug_synchronous)
     {
         HIPCUB_DETAIL_RUNTIME_LOG_DEBUG_SYNCHRONOUS();
         return Encode(d_temp_storage,